requirements.txt
blinker
click
Flask
itsdangerous
Jinja2
MarkupSafe
numpy
pillow
Werkzeug
